@@ -1,7 +1,7 @@
 package github.javaguide.springsecurityjwtguide.security.filter;
 
 import github.javaguide.springsecurityjwtguide.security.common.constants.SecurityConstants;
-import github.javaguide.springsecurityjwtguide.security.common.utils.JwtTokenUtils;
+import github.javaguide.springsecurityjwtguide.security.common.utils.JwtTokenUtil;
 import io.jsonwebtoken.JwtException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.redis.core.StringRedisTemplate;
@@ -10,10 +10,10 @@ import org.springframework.security.authentication.UsernamePasswordAuthenticatio
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.web.authentication.www.BasicAuthenticationFilter;
 
-import javax.servlet.FilterChain;
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
+import jakarta.servlet.FilterChain;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
 /**
@@ -37,24 +37,46 @@ public class JwtAuthorizationFilter extends BasicAuthenticationFilter {
 
         String token = request.getHeader(SecurityConstants.TOKEN_HEADER);
         if (token == null || !token.startsWith(SecurityConstants.TOKEN_PREFIX)) {
+            log.debug("No token found in request header or invalid format");
             SecurityContextHolder.clearContext();
             chain.doFilter(request, response);
             return;
         }
+        
         String tokenValue = token.replace(SecurityConstants.TOKEN_PREFIX, "");
         UsernamePasswordAuthenticationToken authentication = null;
         try {
-            String previousToken = stringRedisTemplate.opsForValue().get(JwtTokenUtils.getId(tokenValue));
-            if (!token.equals(previousToken)) {
+            String previousToken = stringRedisTemplate.opsForValue().get(JwtTokenUtil.getId(tokenValue));
+            if (previousToken == null) {
+                log.warn("Token not found in Redis - user may be logged out or token expired");
                 SecurityContextHolder.clearContext();
                 chain.doFilter(request, response);
                 return;
             }
-            authentication = JwtTokenUtils.getAuthentication(tokenValue);
+            
+            if (!token.equals(previousToken)) {
+                log.warn("Token in request doesn't match stored token - possible token reuse attempt");
+                SecurityContextHolder.clearContext();
+                chain.doFilter(request, response);
+                return;
+            }
+            
+            authentication = JwtTokenUtil.getAuthentication(tokenValue);
+            if (authentication != null && authentication.getPrincipal() != null) {
+                log.debug("Successfully authenticated user: {}", authentication.getPrincipal());
+                SecurityContextHolder.getContext().setAuthentication(authentication);
+            } else {
+                log.warn("Could not create authentication from token");
+                SecurityContextHolder.clearContext();
+            }
         } catch (JwtException e) {
-            logger.error("Invalid jwt : " + e.getMessage());
+            log.error("Invalid JWT token: {}", e.getMessage());
+            SecurityContextHolder.clearContext();
+        } catch (Exception e) {
+            log.error("Error processing authentication: {}", e.getMessage());
+            SecurityContextHolder.clearContext();
         }
-        SecurityContextHolder.getContext().setAuthentication(authentication);
+        
         chain.doFilter(request, response);
     }
 }
