@@ -9,7 +9,7 @@ import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseBody;
 
-import javax.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletRequest;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -24,8 +24,10 @@ public class GlobalExceptionHandler {
 
     @ExceptionHandler(BaseException.class)
     public ResponseEntity<ErrorResponse> handleBaseException(BaseException ex, HttpServletRequest request) {
-        ErrorResponse errorResponse = new ErrorResponse(ex, request.getRequestURI());
-        log.error("occur BaseException:" + errorResponse.toString());
+        ErrorResponse errorResponse = new ErrorResponse(ex.getErrorCode().getMessage(), 
+                                                       ex.getErrorCode().getStatus().value(), 
+                                                       request.getRequestURI());
+        log.error("occur BaseException:" + errorResponse);
         return ResponseEntity.status(ex.getErrorCode().getStatus()).body(errorResponse);
     }
 
@@ -34,31 +36,46 @@ public class GlobalExceptionHandler {
      */
     @ExceptionHandler(MethodArgumentNotValidException.class)
     public ResponseEntity<ErrorResponse> handleMethodArgumentNotValidException(MethodArgumentNotValidException ex, HttpServletRequest request) {
-        Map<String, Object> errors = new HashMap<>(8);
+        StringBuilder errorMessage = new StringBuilder("Validation failed: ");
         ex.getBindingResult().getAllErrors().forEach(error -> {
             String fieldName = ((FieldError) error).getField();
-            String errorMessage = error.getDefaultMessage();
-            errors.put(fieldName, errorMessage);
+            String message = error.getDefaultMessage();
+            errorMessage.append(fieldName).append(" ").append(message).append(", ");
         });
-        ErrorResponse errorResponse = new ErrorResponse(ErrorCode.METHOD_ARGUMENT_NOT_VALID, request.getRequestURI(), errors);
-        log.error("occur MethodArgumentNotValidException:" + errorResponse.toString());
+        
+        // Remove trailing comma and space
+        String finalMessage = errorMessage.substring(0, errorMessage.length() - 2);
+        
+        ErrorResponse errorResponse = new ErrorResponse(
+            finalMessage,
+            HttpStatus.BAD_REQUEST.value(),
+            request.getRequestURI()
+        );
+        
+        log.error("occur MethodArgumentNotValidException:" + errorResponse);
         return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(errorResponse);
     }
 
     @ExceptionHandler(value = UserNameAlreadyExistException.class)
     public ResponseEntity<ErrorResponse> handleUserNameAlreadyExistException(UserNameAlreadyExistException ex, HttpServletRequest request) {
-        ErrorResponse errorResponse = new ErrorResponse(ex, request.getRequestURI());
-        log.error("occur UserNameAlreadyExistException:" + errorResponse.toString());
+        ErrorResponse errorResponse = new ErrorResponse(
+            ex.getErrorCode().getMessage(),
+            ex.getErrorCode().getStatus().value(),
+            request.getRequestURI()
+        );
+        log.error("occur UserNameAlreadyExistException:" + errorResponse);
 
         return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(errorResponse);
     }
 
     @ExceptionHandler(value = {RoleNotFoundException.class, UserNameNotFoundException.class})
     public ResponseEntity<ErrorResponse> handleUserNotFoundException(BaseException ex, HttpServletRequest request) {
-        ErrorResponse errorResponse = new ErrorResponse(ex, request.getRequestURI());
-        log.error("occur ResourceNotFoundException:" + errorResponse.toString());
+        ErrorResponse errorResponse = new ErrorResponse(
+            ex.getErrorCode().getMessage(),
+            ex.getErrorCode().getStatus().value(),
+            request.getRequestURI()
+        );
+        log.error("occur ResourceNotFoundException:" + errorResponse);
         return ResponseEntity.status(HttpStatus.NOT_FOUND).body(errorResponse);
     }
-
-
 }
