@@ -1,6 +1,7 @@
 package github.javaguide.springsecurityjwtguide.system.validator;
 
-import javax.validation.Constraint;
+import jakarta.validation.Constraint;
+import jakarta.validation.Payload;
 import java.lang.annotation.Documented;
 import java.lang.annotation.Retention;
 import java.lang.annotation.Target;
@@ -10,13 +11,13 @@ import static java.lang.annotation.ElementType.PARAMETER;
 import static java.lang.annotation.RetentionPolicy.RUNTIME;
 
 @Documented
-@Constraint(validatedBy = FullNameValidator.class)
+@Constraint(validatedBy = {FullNameValidator.class})
 @Target({PARAMETER, FIELD})
 @Retention(RUNTIME)
 public @interface FullName {
     String message() default "姓名格式错误";
 
-    Class[] groups() default {};
+    Class<?>[] groups() default {};
 
-    Class[] payload() default {};
+    Class<? extends Payload>[] payload() default {};
 }
