@@ -1,8 +1,8 @@
 package github.javaguide.springsecurityjwtguide.system.validator;
 
 
-import javax.validation.ConstraintValidator;
-import javax.validation.ConstraintValidatorContext;
+import jakarta.validation.ConstraintValidator;
+import jakarta.validation.ConstraintValidatorContext;
 
 public class FullNameValidator implements ConstraintValidator<FullName, String> {
     /**
