@@ -4,8 +4,6 @@ import github.javaguide.springsecurityjwtguide.system.service.UserService;
 import github.javaguide.springsecurityjwtguide.system.web.representation.UserRepresentation;
 import github.javaguide.springsecurityjwtguide.system.web.request.UserRegisterRequest;
 import github.javaguide.springsecurityjwtguide.system.web.request.UserUpdateRequest;
-import io.swagger.annotations.Api;
-import io.swagger.annotations.ApiOperation;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
@@ -20,21 +18,22 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-import javax.validation.Valid;
+import jakarta.validation.Valid;
 
 /**
  * @author shuang.kou
  */
 @RestController
-@RequiredArgsConstructor(onConstructor = @__(@Autowired))
 @RequestMapping("/users")
-@Api(tags = "用户")
 public class UserController {
 
     private final UserService userService;
 
+    public UserController(UserService userService) {
+        this.userService = userService;
+    }
+
     @PostMapping("/sign-up")
-    @ApiOperation("用户注册")
     public ResponseEntity<Void> signUp(@RequestBody @Valid UserRegisterRequest userRegisterRequest) {
         userService.save(userRegisterRequest);
         return ResponseEntity.ok().build();
@@ -42,15 +41,16 @@ public class UserController {
 
     @GetMapping
     @PreAuthorize("hasAnyRole('ROLE_USER','ROLE_MANAGER','ROLE_ADMIN')")
-    @ApiOperation("获取所有用户的信息（分页）")
-    public ResponseEntity<Page<UserRepresentation>> getAllUser(@RequestParam(value = "pageNum", defaultValue = "0") int pageNum, @RequestParam(value = "pageSize", defaultValue = "10") int pageSize) {
+    public ResponseEntity<Page<UserRepresentation>> getAllUser(
+            @RequestParam(value = "pageNum", defaultValue = "0") int pageNum, 
+            @RequestParam(value = "pageSize", defaultValue = "10") int pageSize) {
+        
         Page<UserRepresentation> allUser = userService.getAll(pageNum, pageSize);
         return ResponseEntity.ok().body(allUser);
     }
 
     @PutMapping
     @PreAuthorize("hasAnyRole('ROLE_ADMIN')")
-    @ApiOperation("更新用户")
     public ResponseEntity<Void> update(@RequestBody @Valid UserUpdateRequest userUpdateRequest) {
         userService.update(userUpdateRequest);
         return ResponseEntity.ok().build();
@@ -58,7 +58,6 @@ public class UserController {
 
     @DeleteMapping
     @PreAuthorize("hasAnyRole('ROLE_ADMIN')")
-    @ApiOperation("根据用户名删除用户")
     public ResponseEntity<Void> deleteUserByUserName(@RequestParam("username") String username) {
         userService.delete(username);
         return ResponseEntity.ok().build();
