@@ -1,13 +1,12 @@
 package github.javaguide.springsecurityjwtguide.system.web.request;
 
-
 import github.javaguide.springsecurityjwtguide.system.entity.User;
 import github.javaguide.springsecurityjwtguide.system.validator.FullName;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import javax.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotBlank;
 
 /**
  * @author shuang.kou
@@ -25,8 +24,10 @@ public class UserRegisterRequest {
     private String fullName;
 
     public User toUser() {
-        return User.builder().fullName(this.getFullName())
+        return User.builder()
+                .fullName(this.getFullName())
                 .userName(this.getUserName())
-                .enabled(true).build();
+                .enabled(true)
+                .build();
     }
 }
