@@ -1,11 +1,16 @@
 package github.javaguide.springsecurityjwtguide;
 
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
 
-@RunWith(SpringRunner.class)
+@ExtendWith(SpringExtension.class)
 @SpringBootTest
-public class SpringSecurityJwtGuideApplicationTests {
+class SpringSecurityJwtGuideApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
 
 }
