@@ -3,13 +3,11 @@ package github.javaguide.springsecurityjwtguide.security.common.utils;
 import github.javaguide.springsecurityjwtguide.security.common.constants.SecurityConstants;
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.SignatureAlgorithm;
 import io.jsonwebtoken.security.Keys;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 
 import javax.crypto.SecretKey;
-import javax.xml.bind.DatatypeConverter;
 import java.util.Arrays;
 import java.util.Date;
 import java.util.List;
@@ -19,29 +17,32 @@ import java.util.stream.Collectors;
  * @author shuang.kou
  * @description JWT工具类
  */
-public class JwtTokenUtils {
-
+public class JwtTokenUtil {
 
     /**
      * 生成足够的安全随机密钥，以适合符合规范的签名
      */
-    private static final byte[] API_KEY_SECRET_BYTES = DatatypeConverter.parseBase64Binary(SecurityConstants.JWT_SECRET_KEY);
+    private static final byte[] API_KEY_SECRET_BYTES = SecurityConstants.JWT_SECRET_KEY.getBytes();
     private static final SecretKey SECRET_KEY = Keys.hmacShaKeyFor(API_KEY_SECRET_BYTES);
 
     public static String createToken(String username, String id, List<String> roles, boolean isRememberMe) {
         long expiration = isRememberMe ? SecurityConstants.EXPIRATION_REMEMBER : SecurityConstants.EXPIRATION;
         final Date createdDate = new Date();
         final Date expirationDate = new Date(createdDate.getTime() + expiration * 1000);
+        
         String tokenPrefix = Jwts.builder()
-                .setHeaderParam("type", SecurityConstants.TOKEN_TYPE)
-                .signWith(SECRET_KEY, SignatureAlgorithm.HS256)
+                .header()
+                .type(SecurityConstants.TOKEN_TYPE)
+                .and()
+                .issuer("SnailClimb")
+                .subject(username)
+                .issuedAt(createdDate)
+                .expiration(expirationDate)
+                .id(id)
                 .claim(SecurityConstants.ROLE_CLAIMS, String.join(",", roles))
-                .setId(id)
-                .setIssuer("SnailClimb")
-                .setIssuedAt(createdDate)
-                .setSubject(username)
-                .setExpiration(expirationDate)
+                .signWith(SECRET_KEY)
                 .compact();
+                
         return SecurityConstants.TOKEN_PREFIX + tokenPrefix; // 添加 token 前缀 "Bearer ";
     }
 
@@ -66,9 +67,9 @@ public class JwtTokenUtils {
 
     private static Claims getClaims(String token) {
         return Jwts.parser()
-                .setSigningKey(SECRET_KEY)
-                .parseClaimsJws(token)
-                .getBody();
+                .verifyWith(SECRET_KEY)
+                .build()
+                .parseSignedClaims(token)
+                .getPayload();
     }
-
 }
